Updated Functioning Vehicle Exe
There was a primary issue with the vehicle exe code prior to the demo in
that when making frequent requests to the API for the vehicle state, the
API had not been thread safe in the request and ultimately returned a
value that would break the cycle of the aircraft performing the guided
mission. This however is corrected by creating a callback and handling
the udpate inside there. The next commit shall remove the guided
activation per mode change (This was done so testing could take place
without requiring a GUI to fire the aircraft state)