Base Vehicle Class Structure
There now exists a data object that sets up the vehicle data state
transmission. The generic object is a vehice_data_state which creates a
vehicle_data object which can be made up of varying protocols data
structure. This structure can then be used to pass the needed mapped
item from the module to core or vice-versa. It will be then show that
the object can be used to help populate the relevant vehicle object and
then use the virtual functions in a get/set methodology to acquire the
desired information.