Start server on module addition
Still not working. Starting a TCP request (in this case using telnet)
can't talk to the Mace application. Testing with a test Qt console
application that starts the Ground Station tcp server works using telnet
to initiate a TCP request. See serverTest application for an example of
a working TCP server.