UDP Sender Port
TODO: The listenForPort function is blocking, I need to move it to a
thread so that it can run in the background. @MadisonBlake if you have
time, could you look at the TODO-PAT comments in
module_vehicle_mavlink.cpp and suggest a course of action for moving
that code to a separate thread?