Vehicle takeoff dialog
When clicking takeoff, a dialog appears allowing a user to change
takeoff altitude. If no vehicle is selected, a dropdown is there to
select the vehicle. You can also set the default takeoff altitude in the
settings menu. TODO: Not set the default takeoff altitude when actually
commanding a takeoff.