Update vehicle position function added
When a vehicle position needs updating, call this function. Pass in "true" as the recompute voronoi flag to recompute, "false" to simply add to our vehicle map. Note that not recomputing will not update the number of cells or their positions. Passing in true will update the number of cells and their corresponding sites with the last known vehicle position.