Beginning Mission Controller Interface
The controller interface now is spawned on a vehicle data objects own thread to self-manage comms and dropouts when exchanging mission items with an aircraft. The interface has been establishing back to the data object. Now we need to connect the receiving and transmitting functions. We need to somehow store the data object in a classess type object to call for transmitting once the timer expires.