GUI heartbeat handled properly
When the heartbeat "last heard from" value is greater than 30, we request a sync of that vehicle's data. If it get's above 60 seconds, we delete it from the DB and clear it's data from the GUI. On next heartbeat, the vehicle should be repopulated as if a new vehicle came online.